minor organizational mod and clarification
