+++ dvid_point_cloud/sampling.py
@@ -7,12 +7,38 @@ import numpy as np
 import pandas as pd
 
 from .client import DVIDClient
-from .parse import (
-    parse_rles, 
-    vectorized_sample_from_rles
-)
+from .parse import parse_rles
 
 logger = logging.getLogger(__name__)
+
+def vectorized_sample_from_rles(starts_zyx: np.ndarray, lengths: np.ndarray, 
+                               num_points: int) -> np.ndarray:
+    """
+    Generate a point cloud sample from run-length encoded data using vectorized operations.
+    
+    Args:
+        starts_zyx: Array of shape (N, 3) with the ZYX start coordinates of each run
+        lengths: Array of shape (N,) with the length of each run
+        num_points: Number of points to sample
+        
+    Returns:
+        Array of shape (num_points, 3) with the ZYX coordinates of sampled points
+    """
+    # Sample rows with probability proportional to run length
+    chosen_rows = np.random.choice(
+        len(starts_zyx),
+        num_points,
+        replace=True,
+        p=lengths / lengths.sum()
+    )
+    
+    # Get the start coordinates for each sampled row
+    points_zyx = starts_zyx[chosen_rows].copy()
+    
+    # Add random offset in the X dimension (Z in ZYX coordinates)
+    points_zyx[:, 2] += np.random.randint(0, lengths[chosen_rows])
+    
+    return points_zyx
 
 def uniform_sample(server: str, uuid: str, label_id: int, 
                   density_or_count: Union[float, int],
@@ -22,6 +48,7 @@ def uniform_sample(server: str, uuid: str, label_id: int,
                   output_format: str = "xyz") -> Union[np.ndarray, pd.DataFrame]:
     """
     Generate a uniform point cloud sample from a DVID label using a vectorized approach.
+    Note that there's no guarantee that all points will be unique though it's likely.
     
     Args:
         server: DVID server URL
